Appled code of Codepen to my app, and Accordion style is clean adjust to my app.
