+++ frontend/src/views/goal/note/Note.vue
@@ -11,24 +11,20 @@
           </v-col>
         </v-row>
         <v-row> </v-row>
+
         <v-row>
           <v-col cols="3"></v-col>
-          <v-col cols="7">
-            <p class="writing-text content-splitter">
-              <!-- <template>
-                <v-expansion-panels>
-                  <v-expansion-panel v-for="(item, i) in 5" :key="i">
-                    <v-expansion-panel-header>Item</v-expansion-panel-header>
-                    <v-expansion-panel-content>
-                      Lorem ipsum dolor sit amet, consectetur adipiscing elit,
-                      sed do eiusmod tempor incididunt ut labore et dolore magna
-                      aliqua. Ut enim ad minim veniam, quis nostrud exercitation
-                      ullamco laboris nisi ut aliquip ex ea commodo consequat.
-                    </v-expansion-panel-content>
-                  </v-expansion-panel>
-                </v-expansion-panels>
-              </template> -->
-            </p>
+          <v-col cols="6">
+            <div v-for="note in noteList" :key="note.id">
+              <accordion>
+                <div slot="header">{{ note.title }}</div>
+                <div slot="body">{{ note.main }}</div>
+              </accordion>
+              <accordion theme="purple">
+                <div slot="header">TIPS</div>
+                Purple accordion
+              </accordion>
+            </div>
           </v-col>
         </v-row>
       </v-container>
@@ -37,11 +33,23 @@
 </template>
 
 <script>
+import accordion from "./Accordion";
+
 export default {
+  components: {
+    accordion,
+  },
   data: function() {
     return {
       start_date: "2019-01-06",
       deadline: "2019-05-08",
+      noteList: [
+        {
+          main:
+            "<h1>React</h1>React and Vue share many similarities. They both:<p>・utilize a virtual DOM</p><p>・provide reactive and composable view components</p><p>・maintain focus in the</p>core library, with concerns such as routing 　and global state management handled by companion libraries Being so similar in scope, we’ve put more time into fine-tuning this comparison than any other. We want to ensure not only technical accuracy, but also balance. We point out where React  outshines Vue, for example in the richness of their ecosystem and abundance of their custom renderers. With that said, it’s inevitable that the comparison would appear biased towards Vue to some React users, as many of the subjects explored are to some extent subjective. We acknowledge the existence of varying technical taste, and this comparison primarily aims to outline the reasons why Vue could potentially be a better fit if your preferences happen to coincide with ours. Some of the sections below may also be slightly outdated due to recent updates in React 16+, and we are planning to work with the React community to revamp this section in the near future.",
+          title: "Hello World",
+        },
+      ],
     };
   },
 };
